+++ src/main/java/io/github/togar2/pvp/potion/effect/GlowingPotionEffect.java
@@ -10,12 +10,12 @@ public class GlowingPotionEffect extends CombatPotionEffect {
 	}
 	
 	@Override
-	public void onApplied(LivingEntity entity, byte amplifier, CombatVersion version) {
+	public void onApplied(LivingEntity entity, int amplifier, CombatVersion version) {
 		entity.setGlowing(true);
 	}
 	
 	@Override
-	public void onRemoved(LivingEntity entity, byte amplifier, CombatVersion version) {
+	public void onRemoved(LivingEntity entity, int amplifier, CombatVersion version) {
 		entity.setGlowing(false);
 	}
 }
